saas_finder/requirements.txt
langchain
langchain-tavily
tavily-python
python-dotenv
